Extract types to types.ts
