+++ src/bindings/js/node/lib/index.ts
@@ -1,7 +1,9 @@
-const addon: ovNodeModule = require('../build/Release/ov_node_addon.node');
-import loadModel, { ovNodeModule } from './node-model';
+import loadModel from './node-model';
 import createModule from 'openvinojs-common'
 
+import type { ovNodeModule } from './types';
+
+const addon: ovNodeModule = require('../build/Release/ov_node_addon.node');
 
 export default
   createModule('node', loadModel, getVersionString, getDescriptionString);
